[Refactor] Extract localization common functionality (#589)
