+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Tests/LocalizationProviderMocking.swift
@@ -7,6 +7,7 @@ import Foundation
 @testable import AzureCommunicationUICalling
 
 class LocalizationProviderMocking: LocalizationProviderProtocol {
+
     var isApplyCalled: Bool = false
     var isGetLocalizedStringCalled: Bool = false
     var isGetLocalizedStringWithArgsCalled: Bool = false
@@ -19,12 +20,13 @@ class LocalizationProviderMocking: LocalizationProviderProtocol {
         isApplyCalled = true
     }
 
-    func getLocalizedString(_ key: LocalizationKey) -> String {
+    func getLocalizedString<Key>(_ key: Key) -> String
+    where Key: RawRepresentable, Key.RawValue == String {
         isGetLocalizedStringCalled = true
         return key.rawValue
     }
 
-    func getLocalizedString(_ key: LocalizationKey, _ args: CVarArg...) -> String {
+    func getLocalizedString<Key>(_ key: Key, _ args: CVarArg...) -> String where Key: RawRepresentable, Key.RawValue == String {
         isGetLocalizedStringWithArgsCalled = true
         return key.rawValue
     }
