+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Tests/Presentation/Calling/OnHoldOverlayViewModelTests.swift
@@ -8,11 +8,13 @@ import XCTest
 @testable import AzureCommunicationUICalling
 
 class OnHoldOverlayViewModelTests: XCTestCase {
-    var localizationProvider: LocalizationProviderMocking!
+    var localizationProviderMock: LocalizationProviderMocking!
+    var localizationProvider: LocalizationProviderProtocol!
 
     override func setUp() {
         super.setUp()
-        localizationProvider = LocalizationProviderMocking()
+        localizationProviderMock = LocalizationProviderMocking()
+        localizationProvider = localizationProviderMock
     }
 
     override func tearDown() {
@@ -28,14 +30,14 @@ class OnHoldOverlayViewModelTests: XCTestCase {
     func test_onHoldOverlayViewModel_displays_title_from_LocalizationMocking() {
         let sut = makeSUTLocalizationMocking()
         XCTAssertEqual(sut.title, "AzureCommunicationUICalling.OnHoldView.Text.OnHold")
-        XCTAssertTrue(localizationProvider.isGetLocalizedStringCalled)
+        XCTAssertTrue(localizationProviderMock.isGetLocalizedStringCalled)
     }
 
     func test_onHoldOverlayViewModel_errorInfoMode_hasTitle_and_subtitle_from_LocalizationMocking() {
         let sut = makeSUTLocalizationMocking()
         XCTAssertNotNil(sut.errorInfoViewModel)
-        XCTAssertEqual(sut.errorInfoViewModel?.title, localizationProvider.getLocalizedString(.snackBarErrorOnHoldTitle))
-        XCTAssertEqual(sut.errorInfoViewModel?.subtitle, localizationProvider.getLocalizedString(.snackBarErrorOnHoldSubtitle))
+        XCTAssertEqual(sut.errorInfoViewModel?.title, localizationProvider.getLocalizedString(LocalizationKey.snackBarErrorOnHoldTitle))
+        XCTAssertEqual(sut.errorInfoViewModel?.subtitle, localizationProvider.getLocalizedString(LocalizationKey.snackBarErrorOnHoldSubtitle))
     }
 
     func test_onHoldOverlayViewModel_tapActionPeformed_when_actionButton_isTapped() {
@@ -50,7 +52,7 @@ class OnHoldOverlayViewModelTests: XCTestCase {
 }
 
 extension OnHoldOverlayViewModelTests {
-    func makeSUT(localizationProvider: LocalizationProviderMocking? = nil) -> OnHoldOverlayViewModel {
+    func makeSUT(localizationProvider: LocalizationProviderProtocol? = nil) -> OnHoldOverlayViewModel {
         let logger = LoggerMocking()
         let storeFactory = StoreFactoryMocking()
         let factoryMocking = CompositeViewModelFactoryMocking(logger: logger, store: storeFactory.store)
